Add files via upload
There is one minor flaw in the code. It is supposed to break out of the main loop if all given input values are equal to zero, however the case seems to be occuring if 2 of the values are equal to 0 too. Nonetheless though, it works properly for proper date inputs :) I made slight adjustments from the main code and used if/do-while structure, original code had only one while loop which kept recurring as long as all 3 values were not equal to 0. It had the same exit issue too.